Fixed compiler warning

+++ emulator/src/FakePrint.cpp
@@ -59,7 +59,7 @@ size_t Print::printf(const char *format, ...)
         va_end(arg);
         return 0;
     };
-    if(len >= sizeof(loc_buf)){
+    if(len >= (long) sizeof(loc_buf)){
         temp = (char*) malloc(len+1);
         if(temp == NULL) {
             va_end(arg);
